JDBCConnection-methods {RJDBC}	R Documentation

Methods for the class ‘JDBCConnect’ in Package ‘RJDBC’

Description

Methods for the class ‘JDBCConnection’ in Package ‘RJDBC’.

dbSendQuery and dbSendUpdate submit a SQL query to the database. The difference between the two is only that dbSendUpdate is used with DBML queries and thus doesn't return any result set.

dbGetTables and dbGetFields are similar to dbListTables and dbListFields but the result is a data frame with all available details (whereas the latter return only a character vector of the names).

Usage

dbSendUpdate (conn, statement, ...)
dbGetTables (conn, ...)
dbGetFields (conn, ...)

Arguments

conn
connection object

statement
SQL statement to execute

...
additional arguments to prepared statement substituted for "?"

Methods

dbCommit

signature(conn = "JDBCConnection", ...)

dbDataType

signature(dbObj = "JDBCConnection", obj = "ANY", ...)

dbDisconnect

signature(conn = "JDBCConnection", ...)

dbExistsTable

signature(conn = "JDBCConnection", name = "character", ...)

dbGetException

signature(conn = "JDBCConnection", ...)

dbGetFields

signature(conn = "JDBCConnection", ...)

dbGetInfo

signature(conn = "JDBCConnection", ...)

dbGetQuery

signature(conn = "JDBCConnection", statement = "character", ...)

dbGetTables

signature(conn = "JDBCConnection", ...)

dbListFields

signature(conn = "JDBCConnection", ...)

dbListResults

signature(conn = "JDBCConnection", ...)

dbListTables

signature(conn = "JDBCConnection", ...)

dbReadTable

signature(conn = "JDBCConnection", ...)

dbRemoveTable

signature(conn = "JDBCConnection", ...)

dbRollback

signature(conn = "JDBCConnection", ...)

dbSendQuery

signature(conn = "JDBCConnection", statement = "character", ...)

dbSendUpdate

signature(conn = "JDBCConnection", statement = "character", ...)

dbWriteTable

signature(conn = "JDBCConnection", ...)

[Package RJDBC version 0.2-7.1 Index]