JDBCDriver-class {RJDBC}	R Documentation

Class "JDBCDriver"

Description

A DBI driver that uses any JDBC driver to access databases.

Generators

Objects can be created by calls to JDBC or dbDriver.

Slots

identifier.quote:

Quote character to use for identifiers in automatically generated SQL statements or NA if quoted identifiers are not supported by the back-end.

jdrv:

Java object reference to an instance of the driver if the driver can be instantiated by a default constructor. This object is only used as a fall-back when the driver manager fails to find a driver.

Extends

Class "DBIDriver-class", directly. Class "DBIObject-class", by class "DBIDriver", distance 2.

Methods

No methods defined with class "JDBCDriver" in the signature.

See Also

JDBC

[Package RJDBC version 0.2-7.1 Index]