JDBCDriver-methods {RJDBC}	R Documentation

Methods for the class JDBCDriver in Package ‘RJDBC’

Description

Methods for the class ‘JDBCDriver’ in Package ‘RJDBC’.

Most prominent method is dbConnect, it creates a new JDBC connection using the specified driver. Due to the fact that the actual JDBC driver is selected in the URL, the JDBCDriver object itself has little significance and is not used to determine the connection type.

dbListConnections always return NULL with a warning, because JDBC connecteions are not tracked.

dbGetInfo returns very basic information, because the JDBC driver is not loaded until a connection is created.

dbUnloadDriver is a no-op in the current implementation, because drivers are never removed from the JVM.

Methods

dbConnect

signature(drv = "JDBCDriver", ...)

dbListConnections

signature(drv = "JDBCDriver", ...)

dbGetInfo

signature(drv = "JDBCDriver", ...)

dbUnloadDriver

signature(drv = "JDBCDriver", ...)

[Package RJDBC version 0.2-7.1 Index]