JDBCResult-methods {RJDBC}	R Documentation

Methods for the class JDBCResult in Package ‘RJDBC’ ~~

Description

Methods for the class ‘JDBCResult’ in Package ‘RJDBC’.

fetch retrieves the content of the result set in the form of a data frame. If n is -1 then the current implementation fetches 32k rows first and then (if not sufficient) continues with chunks of 512k rows, appending them. If the size of the result set is known in advance, it is most efficient to set n to that size.

Additional argument block can be used to inform the driver that pre-fetching of a certain block of records is desirable (see setFetchSize() in JDBC) leading to possibly faster pulls of large queries. The default is to pre-fetch 2048 records. Note that some databases (like Oracle) don't support a fetch size of more than 32767. If set to NA or anything less than 1 then the fetch size is not changed. This is only a hint, drivers are free to ignore it.

dbClearResult releases the result set.

dbColumnInfo returns basic information about the columns (fields) in the result set.

dbGetInfo returns an empty list.

dbListResults returns an empty list and warns that JDBC doesn't track results.

Methods

fetch

signature(res = "JDBCResult", ...)

dbClearResult

signature(res = "JDBCResult", ...)

dbColumnInfo

signature(res = "JDBCResult", ...)

dbGetInfo

signature(res = "JDBCResult", ...)

dbListResults

signature(res = "JDBCResult", ...)

[Package RJDBC version 0.2-7.1 Index]